Add files via upload
A model that learns to invert STFT representations is built and trained. The model can operate on either the original acoustic signal or a combination of the encoder and decoder. An acoustic waveform is produced as the output. To assess the model's performance, visual and auditory demonstrations are provided. The NSynth dataset is employed. A ReadMe is provided for further details.